+++ mycodo/widgets/widget_measurement.py
@@ -186,7 +186,7 @@ WIDGET_INFORMATION = {
   
   {%- for each_function in function if each_function.unique_id == device_id and measurement_id in device_measurements_dict -%}
 
-    <span style="font-size: {{widget_options['font_em_value']}}em" id="value-{{each_widget.unique_id}}"></span><span style="font-size: {{widget_options['font_em_unit']}}em">
+    <span class="widget-measurement-value" style="font-size: {{widget_options['font_em_value']}}em" id="value-{{each_widget.unique_id}}"></span><span style="font-size: {{widget_options['font_em_unit']}}em">
         {%- if dict_measure_units[measurement_id] in dict_units and
                dict_units[dict_measure_units[measurement_id]]['unit'] and
                widget_options['enable_unit'] -%}
@@ -220,7 +220,7 @@ WIDGET_INFORMATION = {
 
   {%- for each_output in output  if each_output.unique_id == device_id and measurement_id in device_measurements_dict -%}
 
-    <span style="font-size: {{widget_options['font_em_value']}}em" id="value-{{each_widget.unique_id}}"></span><span style="font-size: {{widget_options['font_em_unit']}}em">
+    <span class="widget-measurement-value" style="font-size: {{widget_options['font_em_value']}}em" id="value-{{each_widget.unique_id}}"></span><span style="font-size: {{widget_options['font_em_unit']}}em">
         {%- if dict_measure_units[measurement_id] in dict_units and
                dict_units[dict_measure_units[measurement_id]]['unit'] and
                widget_options['enable_unit'] -%}
@@ -254,7 +254,7 @@ WIDGET_INFORMATION = {
 
   {%- for each_pid in pid  if each_pid.unique_id == device_id and measurement_id in device_measurements_dict -%}
 
-    <span style="font-size: {{widget_options['font_em_value']}}em" id="value-{{each_widget.unique_id}}"></span><span style="font-size: {{widget_options['font_em_unit']}}em">
+    <span class="widget-measurement-value" style="font-size: {{widget_options['font_em_value']}}em" id="value-{{each_widget.unique_id}}"></span><span style="font-size: {{widget_options['font_em_unit']}}em">
         {%- if dict_measure_units[measurement_id] in dict_units and
                dict_units[dict_measure_units[measurement_id]]['unit'] and
                widget_options['enable_unit'] -%}